move header parser to own module
